Progress Afrekenen
-Bon en opmerking toevoegen werkt nu
-Je krijgt een bevestiging als het maken van de bon is gelukt, dan ga je terug naar het tafel keuzemenu

+++ DAL/Bon_DAO.cs
@@ -4,7 +4,7 @@ namespace DAL_Chapoo
 {
     public class Bon_DAO : Base
     {
-        public void Db_Insert_Bon(int BestellingID, int TafelID, double Totaalbedrag, int Fooi, string Betaalmethode)
+        public void Db_Insert_Bon(int BestellingID, int TafelID, double Totaalbedrag, float Fooi, string Betaalmethode)
         {
             string query = "INSERT INTO Bon (BestellingID, TafelID, Totaalbedrag, Fooi, Betaalmethode) VALUES (@BestellingID, @TafelID, @Totaalbedrag, @Fooi, @Betaalmethode)";
             SqlParameter[] sqlParameters = new SqlParameter[5];
